requirements.txt
Flask
langchain
langchain-community
langchain-chroma
langchain-deepseek
pypdf2
python-dotenv
requests
chromadb